add expense_category trait
